Commit One
Unfortunately I didn't end up with enough time to complete the project
as I'd have liked to however the business and data layer are both
working well, but would require more work to be scalable. The front end
is not really usable but does show proof of concept on the page
"/EmployeeView". I have used a small amount of CSS to style the page and
have changed the item templates. This is the first time I've used
webforms so took a while getting to grips with the concept.